Fix cat function
There is a slight possibility that `write` does not write the same
amount of byte that we want write to print. This can happen when `write`
is interrupted.

We avoid this issue when we use fread and fwrite from stdio. So use this
one instead of the low-level open/read/write/close system calls.